Move shared function in shared folder
