+++ lib/suivi-pcrs.js
@@ -1,5 +1,3 @@
-import {maxBy} from 'lodash-es'
-
 const API_URL = process.env.NEXT_PUBLIC_URL || 'https://pcrs.beta.gouv.fr'
 
 async function request(url, options) {
@@ -35,22 +33,6 @@ function buildOptions({method, token, body}) {
   }
 
   return options
-}
-
-export function findClosestEtape(etapes) {
-  if (etapes.some(etape => etape.date_debut)) {
-    const now = new Date()
-
-    const filteredLaterSteps = etapes.filter(etape => new Date(etape.date_debut) <= now)
-    if (filteredLaterSteps.length > 0) {
-      return maxBy(filteredLaterSteps, etape => new Date(etape.date_debut))
-    }
-
-    return etapes[0]
-  }
-
-  // When "Etapes" has no date, use last one
-  return etapes[etapes.length - 1]
 }
 
 export async function getProject(id, token) {
