+++ server/projets.js
@@ -3,7 +3,7 @@ import {omit} from 'lodash-es'
 import {nanoid} from 'nanoid'
 import {validateCreation, validateChanges} from '../lib/projets-validator.js'
 import {buildGeometryFromTerritoires, getTerritoiresProperties} from '../lib/territoires.js'
-import {findClosestEtape} from '../lib/suivi-pcrs.js'
+import {findClosestEtape} from '../shared/find-closest-etape.js'
 import mongo from './util/mongo.js'
 
 export function expandProjet(projet) {
